Specify that the work comes from datacite when claiming
Note this is an intended workaround as we only support datacite at present for claiming.
Ideally we would have the claiming part know which RA its working from to register.
This exists because in some cases we can't automatically look up who the RA is, i.e. fake test prefixes