using version other type
Treatise advancing an original point of view resulting from research: a requirement for a doctoral degree. This is only available with API v2.x A paper that is typically based on original research and that gives evidence of the candidate’s mastery both of their own subject and of scholarly method. This is only available with API v3.0+